Add Applicant Skills report
# New features
* Add new sheet, Applicant Skills, with skill data captured during the online application process
* Add bookmark, French-speaking Canadians, that makes selections with the new skill fields to select Canadian applicants reporting at least Basic (A1/A2) French CEFR proficiency level
  * Applicant Address Country: Canada
  * Skill Description: French
  * Skill Proficiency Level: Fair, Good, or Excellent
  * Application Is Submitted: Yes

# Issues fixed
* Add Vizlib Sheet Menu to sheets that appear to be missing the object (the menu is not loading)

# Additional information
As of approximately mid-2020, the online application has used the [Common European Framework of References for Languages (CEFR) standard](https://www.coe.int/en/web/common-european-framework-reference-languages) when asking users for their levels of proficiency in English and French. However, the data is stored using a more generic system with the same number of levels. Below is a mapping table that shows how the level is present to the user versus how it is stored in the database and displayed in Qlik.

| Appearance in application | Appearance in database and Qlik |
| --------------------------- | ----------------------------------|
| Basic (A1/A2)                       | Fair                                                 |
| Independent (B1/B2)          | Good                                               |
| Proficient (C1/C2)               | Excellent                                         |